refactoring of NLP related helper functions
