+++ prospector/processing/commit/SLAB Ideas.md
