+++ prospector/util/similarity.py
